requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-PyMongo==2.3.0
python-dotenv==1.0.0
pymongo==4.5.0
google-generativeai==0.3.2
langchain==0.1.0
langchain-google-genai==0.0.8
faiss-cpu==1.7.4
PyMuPDF==1.23.8
pdfplumber==0.10.3
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
python-multipart==0.0.6
Werkzeug==2.3.7
bcrypt==4.1.2
PyJWT==2.8.0
